// deploy_temp/src/types/dashboard.ts
export interface Vehiculo {
    id: number;
    patente: string;
    marca: string;
    modelo: string;
    anio: number;
    tipo: 'Camioneta' | 'Furgón' | 'Camión' | 'Automóvil';
    kilometraje: number;
    proximaRevisionTecnica: string; // Date
    vencimientoSeguro: string; // Date
    vencimientoPermisoCirculacion: string; // Date
    estado: 'Operativo' | 'En Taller' | 'Fuera de Servicio';
    conductorAsignado?: string;
    observaciones?: string;
    activo: boolean;
}